[CCUBE-1433][SR|JY] Made changes for function that creates validation schema, making it extensible, rearranged the translator codes accordingly & amended the tests accordingly.
